# lib/api-spec/openapi.yaml
openapi: "3.1.0"
info:
  title: Api
  version: "1.0.0"
  description: MANTHANA-SCHOLER — AI-powered thesis and research writing platform for Indian medical scholars
servers:
  - url: /api
paths:
...
  /profile:
    get:
      operationId: getProfile
      summary: Get authenticated user profile
      tags: [profile]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    put:
      operationId: upsertProfile
      summary: Create or update user profile
      tags: [profile]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserProfileInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"

  # ─── DASHBOARD ───────────────────────────────────────────────────────────────
...
  /workspaces/{workspaceId}/sections/{sectionId}/chat:
    get:
      operationId: listChatMessages
      summary: List chat history for a section
      tags: [chat]
      parameters:
        - in: path
          name: workspaceId
          required: true
          schema:
            type: integer
        - in: path
          name: sectionId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatMessage"

    post:
      operationId: sendChatMessage
      summary: Send a message to the AI assistant
      tags: [chat]
      parameters:
        - in: path
          name: workspaceId
          required: true
          schema:
            type: integer
        - in: path
          name: sectionId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatMessageInput"
      responses:
        "200":
          description: AI response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatMessage"
  /workspaces/{workspaceId}/sections/{sectionId}/chat/clear:
    delete:
      operationId: clearChatHistory
      summary: Clear chat history for a section
      tags: [chat]
      parameters:
        - in: path
          name: workspaceId
          required: true
          schema:
            type: integer
        - in: path
          name: sectionId
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Cleared
  /workspaces/{workspaceId}/sections/{sectionId}/generate:
    post:
      operationId: generateSectionContent
      summary: AI-generate draft content for a section
      tags: [chat]
      parameters:
        - in: path
          name: workspaceId
          required: true
          schema:
            type: integer
        - in: path
          name: sectionId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateContentInput"
      responses:
        "200":
          description: Generated content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneratedContent"

  # ─── VAULT ───────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # ─── REFERENCE ───────────────────────────────────────────────────────────
...
    UserProfile:
      type: object
      required: [id, clerkUserId, fullName, email, domain, qualification, onboardingComplete, createdAt]
      properties:
        id:
          type: integer
        clerkUserId:
          type: string
        fullName:
          type: string
        email:
          type: string
        domain:
          type: string
        qualification:
          type: string
        collegeName:
          type: ["string", "null"]
        universityName:
          type: ["string", "null"]
        guideNames:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        state:
          type: ["string", "null"]
        onboardingComplete:
          type: boolean
        createdAt:
          type: string
    UserProfileInput:
      type: object
      required: [fullName, email, domain, qualification]
      properties:
        fullName:
          type: string
        email:
          type: string
        domain:
          type: string
        qualification:
          type: string
        collegeName:
          type: string
        universityName:
          type: string
        guideNames:
          type: string
        city:
          type: string
        state:
          type: string
        onboardingComplete:
          type: boolean

    # ─── DASHBOARD ───────────────────────────────────────────────────────────
...
    ChatMessage:
      type: object
      required: [id, sectionId, role, content, createdAt]
      properties:
        id:
          type: integer
        sectionId:
          type: integer
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
        tokensUsed:
          type: ["integer", "null"]
        createdAt:
          type: string
    ChatMessageInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
          minLength: 1
        includeContext:
          type: boolean
    GenerateContentInput:
      type: object
      required: [prompt]
      properties:
        prompt:
          type: string
        tone:
          type: string
          enum: [academic, concise, detailed, formal]
        wordLimit:
          type: integer
    GeneratedContent:
      type: object
      required: [content, tokensUsed]
      properties:
        content:
          type: string
        tokensUsed:
          type: integer

    # ─── VAULT ───────────────────────────────────────────────────────────────
...